Finished data flow changes (to the point that FFT processing is happening again)

+++ src/test/kotlin/toolkit/monkeyTest/NextFrequencyBin.kt
@@ -5,7 +5,7 @@ import kotlin.random.Random
 
 fun Random.nextFrequencyBin(): FrequencyBin {
     return FrequencyBin(
-        frequency = Random.nextDouble(),
+        frequency = Random.nextFloat(),
         amplitude = Random.nextDouble()
     )
 }
